Fix missing SO equations
The DO-to-SO mapping (for rim number generation) is one-to-many. This caused the script to stop at the first instance of DO_xxxx, even though the SO number did not match. The bug has been resolved by sending in the soName along with the DO_xxxx name during rim generation method call.